Final touch up : admin page, cart page, detail page

+++ adminpg/delete.php
@@ -2,9 +2,10 @@
 require 'conn.php';
 
 $id = $_GET['id'];
-$conn->query("DELETE FROM barang WHERE id='$id'");
-
-if ($conn) {
+$id2 = $_GET['id2'];
+if ($id2==null) {
+    $conn->query("DELETE FROM barang WHERE id='$id'");
+    if ($conn) {
     echo "<script>
         alert('Data berhasil dihapus!');
         document.location.href = 'index.php';
@@ -15,3 +16,20 @@ if ($conn) {
         document.location.href = 'index.php';
     </script>";
 }
+}
+if ($id==null) {
+    $conn->query("DELETE FROM cart WHERE idbrg='$id2'");
+    if ($conn) {
+    echo "<script>
+        alert('Data berhasil dihapus!');
+        document.location.href = '../user-page/cart.php';
+    </script>";
+} else {
+    echo "<script>
+        alert('Data gagal dihapus!');
+        document.location.href = '../user-page/cart.php';
+    </script>";
+}
+}
+
+
